renamed files used dashes to seperate words
