+++ http-test-clients/menu-routes-test.http
@@ -1,0 +1,181 @@
+### route of the api server
+GET http://localhost:3010/ HTTP/1.1
+
+### Get all menu items
+GET http://localhost:3010/menu/ HTTP/1.1
+
+### get menu item by id
+GET http://localhost:3010/menu/5f0f9a28036e9a473e082a37 HTTP/1.1
+
+### with invalid id
+GET http://localhost:3010/menu/3 HTTP/1.1
+
+### with id that has been deleted
+GET http://localhost:3010/menu/5f0fefe35ae353518e24508f HTTP/1.1
+
+### get menu item of the day -- for home screen
+GET http://localhost:3010/menu/oftheday HTTP/1.1
+
+### create menu item
+POST http://localhost:3010/menu HTTP/1.1
+content-type: application/json
+
+{
+  "title": "pizzas",
+  "description": "Home made",
+  "deliversOn": "2020-07-13T08:08:11+10:00",
+  "mealType": "lunch",
+  "orderStarts": "2020-07-12T10:08:11+10:00",
+  "orderEnds": "2020-07-12T16:08:11+10:00",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### update menu item by id
+PUT http://localhost:3010/menu/5f0f9a28036e9a473e082a37 HTTP/1.1
+content-type: application/json
+
+{
+  "title": "Burger",
+  "description": "Home made",
+  "deliversOn": "2020-07-13T08:08:11+10:00",
+  "mealType": "lunch",
+  "orderStarts": "2020-07-12T10:08:11+10:00",
+  "orderEnds": "2020-07-12T16:08:11+10:00",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### update menu item by id that has been deleted
+PUT http://localhost:3010/menu/5f0fefe35ae353518e24508f HTTP/1.1
+content-type: application/json
+
+{
+  "title": "Burger",
+  "description": "Home made",
+  "deliversOn": "2020-07-15",
+  "mealType": "Dinner",
+  "orderStarts": "2020-07-12",
+  "orderEnds": "2020-07-16",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### update menu item with invalid id
+PUT http://localhost:3010/menu/9 HTTP/1.1
+content-type: application/json
+
+{
+  "title": "Dumpling",
+  "description": "Home made",
+  "deliversOn": "2020-07-15",
+  "mealType": "Dinner",
+  "orderStarts": "2020-07-12",
+  "orderEnds": "2020-07-16",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### delete menu item by id 
+DELETE  http://localhost:3010/menu/5f0f9a28036e9a473e082a37 HTTP/1.1
+content-type: application/json
+
+### delete menu item by id that has been deleted
+DELETE  http://localhost:3010/menu/5f0fefe35ae353518e24508f HTTP/1.1
+content-type: application/json
+
+### delete menu item with invalid id
+DELETE  http://localhost:3010/menu/5 HTTP/1.1
+content-type: application/json
+
+
+### Tests for staging
+### route of the api server
+GET https://homemade-meals-server-dev.herokuapp.com/ HTTP/1.1
+
+### Get all menu items
+GET https://homemade-meals-server-dev.herokuapp.com/menu/ HTTP/1.1
+
+### get menu item by id
+GET https://homemade-meals-server-dev.herokuapp.com/menu/5f11437de481430004f67bff HTTP/1.1
+
+### with invalid id
+GET https://homemade-meals-server-dev.herokuapp.com/menu/3 HTTP/1.1
+
+### with id that has been deleted
+GET https://homemade-meals-server-dev.herokuapp.com/menu/5f0fefe35ae353518e24508f HTTP/1.1
+
+### get menu item of the day -- for home screen
+GET https://homemade-meals-server-dev.herokuapp.com/menu/oftheday HTTP/1.1
+
+### create menu item
+POST https://homemade-meals-server-dev.herokuapp.com/menu HTTP/1.1
+content-type: application/json
+
+{
+  "title": "pizzas",
+  "description": "Home made",
+  "deliversOn": "2020-07-13T08:08:11+10:00",
+  "mealType": "lunch",
+  "orderStarts": "2020-07-12T10:08:11+10:00",
+  "orderEnds": "2020-07-12T16:08:11+10:00",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### update menu item by id
+PUT https://homemade-meals-server-dev.herokuapp.com/menu/5f11437de481430004f67bff HTTP/1.1
+content-type: application/json
+
+{
+  "title": "Burger",
+  "description": "Home made",
+  "deliversOn": "2020-07-13T08:08:11+10:00",
+  "mealType": "lunch",
+  "orderStarts": "2020-07-12T10:08:11+10:00",
+  "orderEnds": "2020-07-12T16:08:11+10:00",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### update menu item by id that has been deleted
+PUT https://homemade-meals-server-dev.herokuapp.com/menu/5f0fefe35ae353518e24508f HTTP/1.1
+content-type: application/json
+
+{
+  "title": "Burger",
+  "description": "Home made",
+  "deliversOn": "2020-07-15",
+  "mealType": "Dinner",
+  "orderStarts": "2020-07-12",
+  "orderEnds": "2020-07-16",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### update menu item with invalid id
+PUT https://homemade-meals-server-dev.herokuapp.com/menu/9 HTTP/1.1
+content-type: application/json
+
+{
+  "title": "Dumpling",
+  "description": "Home made",
+  "deliversOn": "2020-07-15",
+  "mealType": "Dinner",
+  "orderStarts": "2020-07-12",
+  "orderEnds": "2020-07-16",
+  "maxOrders": 20,
+  "cost": 20
+}
+
+### delete menu item by id 
+DELETE  https://homemade-meals-server-dev.herokuapp.com/menu/5f11437de481430004f67bff HTTP/1.1
+content-type: application/json
+
+### delete menu item by id that has been deleted
+DELETE  https://homemade-meals-server-dev.herokuapp.com/menu/5f0fefe35ae353518e24508f HTTP/1.1
+content-type: application/json
+
+### delete menu item with invalid id
+DELETE  https://homemade-meals-server-dev.herokuapp.com/menu/5 HTTP/1.1
+content-type: application/json
